serializer/file_test.go
package serializer

import (
	"github.com/Imomali1/gRPC/pcbook/pb"
	"github.com/Imomali1/gRPC/pcbook/sample"
	"github.com/golang/protobuf/proto"
	"github.com/stretchr/testify/require"
	"testing"
)

func TestFileSerializer(t *testing.T) {
	t.Parallel()

	binaryFile := "../tmp/laptop.bin"
	jsonFile := "../tmp/laptop.json"

	laptop1 := sample.NewLaptop()
	err := WriteProtoBufToBin(laptop1, binaryFile)
	require.NoError(t, err)

	laptop2 := &pb.Laptop{}
	err = ReadProtoBufFromBin(binaryFile, laptop2)
	require.NoError(t, err)
	require.True(t, proto.Equal(laptop1, laptop2))

	err = WriteProtoBufToJSON(laptop1, jsonFile)
	require.NoError(t, err)
}
